Add Euler's pendulum simulation
Euler's Pendulum simulation. Equations of motion is based on Experimental and theoretical investigation of the energy dissipation of a rolling disk during its final stage of motion, R.I.Leine,2008. Magnetic field is generated as the magnetic field of a current loop, the Euler's pendulum is seen as a magnetic cylinder with two surfaces containing magnetic charges, whose surface density is named sigma.